change react select dropdown styling and update components
